requirements.txt
django==3.2.*
psycopg2-binary==2.9.*
python-dotenv==1.1.0
legacy-cgi==2.6.3


